Add Omega, check for Heywood, updated docs
Added Omega Total & Omega Hierachical to the available metrics for
reliability. Switched from pingouin to reliabilipy to this end.

It seems factor_analyzer does not check for Heywood cases, so I added
a check for this and a warning if it is detected.

Updated docs with minor clarifications.

Also updated the pyproject.toml to make optional dependencies actually
optional.

Added requirements.txt and optiona-requirements.txt for pip install.